+++ lib/src/cell_builder.dart
@@ -1,7 +1,8 @@
+import 'package:easy_table/src/row_data.dart';
 import 'package:flutter/widgets.dart';
 
 /// Signature for a function that builds a widget for a given row.
 ///
 /// Used by [EasyTableColumn].
 typedef EasyTableCellBuilder<ROW> = Widget Function(
-    BuildContext context, ROW row, int rowIndex);
+    BuildContext context, RowData<ROW> data);