+++ lib/src/theme/cell_null_color.dart
@@ -1,0 +1,4 @@
+import 'package:flutter/material.dart';
+
+/// Signature for a function that defines a cell null valued color.
+typedef CellNullColor = Color? Function(int rowIndex, bool hovered);